requirements.txt
agno
groq
python-dotenv
browserbase
playwright